+++ jeap-deploymentlog-domain/src/main/java/ch/admin/bit/jeap/deploymentlog/domain/FlowMetricIdentity.java
@@ -1,0 +1,7 @@
+package ch.admin.bit.jeap.deploymentlog.domain;
+
+public record FlowMetricIdentity(String system,
+                                 String component,
+                                 String finalEnvironment,
+                                 FlowType type) {
+}