+++ jeap-deploymentlog-web/src/main/java/ch/admin/bit/jeap/deploymentlog/web/DeploymentLogConfig.java
@@ -4,11 +4,8 @@ import ch.admin.bit.jeap.deploymentlog.docgen.DocumentationGenerator;
 import ch.admin.bit.jeap.deploymentlog.jira.JiraWebClient;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.core.env.Environment;
 
 @AutoConfiguration
 @EnableConfigurationProperties
@@ -17,12 +14,6 @@ import org.springframework.core.env.Environment;
 class DeploymentLogConfig {
 
     private DeploymentLogConfig() {
-    }
-
-    @Bean
-    static BeanPostProcessor prometheusCreatedTimestampDefault(Environment environment) {
-        return new PrometheusCreatedTimestampDefault(environment,
-                Thread.currentThread().getContextClassLoader());
     }
 
 }